+++ HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 1.6.0 (2014-??-??)
 ++++++++++++++++++
+- Added 'temp_dir' argument to Oct2Py constructor (#50)
 - Added 'kill_octave' convenience method to kill zombies (#46)
 - Improved Octave shutdown handling (#45, #46)
 - Added 'oned_as' argument to Oct2Py constructor (#49)